move IMFIABConstraintsChecker to package st2067_201
